rename: 공용 컴포넌트 파일 이동
