+++ components/customSlider.tsx
